refactor: move section heading-color helpers to sectionTitleColors.js
SectioningComponent.js had ~220 lines of helper functions and imports
before the class definition.  Extract them to a dedicated utility module
(src/utils/sectionTitleColors.js) so the component file opens directly
on the class.

Docstrings added to:
  - resolveSectionTitleLightColorSpec (multi-branch resolution with
    inheritance and source-descriptor tracking)
  - deriveSectionTitleDarkModeColor (explains why suggestAccessibleDarkMode-
    ColorAgainst is preferred over deriveAccessibleDarkModeColor)
  - resolveSectionTitleDarkColorSpec (most complex function; four-step
    resolution: own dark → derived from light → inherited → default)
  - shouldEmitSectionTitleColorDiagnostic (dedup logic for nested sections)

No behaviour change; all 33 sectioning tests pass.
