+++ packages/doenetml-worker-javascript/src/components/abstract/SectioningComponent.js
@@ -5,14 +5,6 @@ import {
     setUpVariantSeedAndRng,
 } from "../../utils/variants";
 import { returnStyleDefinitionStateVariables } from "@doenet/utils";
-import {
-    CANVAS_DARK_MODE_COLOR,
-    CANVAS_LIGHT_MODE_COLOR,
-    compositedContrastRatio,
-    invertLightness,
-    suggestAccessibleDarkModeColorAgainst,
-    TEXT_CONTRAST_THRESHOLD,
-} from "@doenet/utils/style";
 import { returnFeedbackDefinitionStateVariables } from "../../utils/feedback";
 import {
     returnScoredSectionAttributes,
@@ -23,229 +15,16 @@ import {
     addChildrenToDynamicChild,
     deleteChildrenFromDynamicChild,
 } from "../../utils/dynamicChildren";
-
-const sectionTitleStateKeys = ["completed", "inProgress", "notStarted"];
-
-function titleStateKeyFromCredit(creditAchieved) {
-    if (creditAchieved === 1) {
-        return "completed";
-    }
-
-    if (creditAchieved > 0) {
-        return "inProgress";
-    }
-
-    return "notStarted";
-}
-
-function resolveSectionTitleLightColorSpec({
-    dependencyValues,
-    usedDefault,
-    ownColorName,
-    parentColors,
-    parentSources,
-    parentIsBoxedOrCollapsible,
-    stateKey,
-}) {
-    if (!usedDefault[ownColorName]) {
-        return {
-            value: dependencyValues[ownColorName],
-            source: {
-                authored: true,
-                colorName: ownColorName,
-                inherited: false,
-                priorBoxedOrCollapsibleAncestor: false,
-            },
-        };
-    }
-
-    const inheritedValue = parentColors?.[stateKey];
-    if (typeof inheritedValue === "string") {
-        return {
-            value: inheritedValue,
-            source: parentSources?.[stateKey]
-                ? {
-                      ...parentSources[stateKey],
-                      inherited: true,
-                      priorBoxedOrCollapsibleAncestor:
-                          parentSources[stateKey]
-                              .priorBoxedOrCollapsibleAncestor ||
-                          parentIsBoxedOrCollapsible,
-                  }
-                : {
-                      authored: false,
-                      colorName: ownColorName,
-                      inherited: true,
-                      priorBoxedOrCollapsibleAncestor:
-                          parentIsBoxedOrCollapsible,
-                  },
-        };
-    }
-
-    return {
-        value: dependencyValues[ownColorName],
-        source: {
-            authored: false,
-            colorName: ownColorName,
-            inherited: false,
-            priorBoxedOrCollapsibleAncestor: false,
-        },
-    };
-}
-
-function deriveSectionTitleDarkModeColor(lightColor) {
-    return suggestAccessibleDarkModeColorAgainst({
-        startColor: invertLightness(lightColor),
-        partnerColor: "#ffffff",
-        channelRole: "background",
-        threshold: TEXT_CONTRAST_THRESHOLD,
-    });
-}
-
-function resolveSectionTitleDarkColorSpec({
-    dependencyValues,
-    usedDefault,
-    ownDarkColorName,
-    ownLightColorName,
-    parentColorsDarkMode,
-    parentSourcesDarkMode,
-    parentIsBoxedOrCollapsible,
-    stateKey,
-}) {
-    if (!usedDefault[ownDarkColorName]) {
-        return {
-            value: dependencyValues[ownDarkColorName],
-            source: {
-                authored: true,
-                colorName: ownDarkColorName,
-                inherited: false,
-                priorBoxedOrCollapsibleAncestor: false,
-            },
-        };
-    }
-
-    if (!usedDefault[ownLightColorName]) {
-        const derivedDarkColor = deriveSectionTitleDarkModeColor(
-            dependencyValues[ownLightColorName],
-        );
-        if (derivedDarkColor !== null) {
-            return {
-                value: derivedDarkColor,
-                source: {
-                    authored: true,
-                    colorName: ownLightColorName,
-                    inherited: false,
-                    priorBoxedOrCollapsibleAncestor: false,
-                },
-            };
-        }
-
-        // CSS variables and other unparseable light-mode colors cannot be
-        // adapted safely in the worker, so fall back to the accessible default
-        // dark-mode color unless the author pins an explicit *DarkMode value.
-        return {
-            value: dependencyValues[ownDarkColorName],
-            source: {
-                authored: false,
-                colorName: ownDarkColorName,
-                inherited: false,
-                priorBoxedOrCollapsibleAncestor: false,
-            },
-        };
-    }
-
-    const inheritedValue = parentColorsDarkMode?.[stateKey];
-    if (typeof inheritedValue === "string") {
-        return {
-            value: inheritedValue,
-            source: parentSourcesDarkMode?.[stateKey]
-                ? {
-                      ...parentSourcesDarkMode[stateKey],
-                      inherited: true,
-                      priorBoxedOrCollapsibleAncestor:
-                          parentSourcesDarkMode[stateKey]
-                              .priorBoxedOrCollapsibleAncestor ||
-                          parentIsBoxedOrCollapsible,
-                  }
-                : {
-                      authored: false,
-                      colorName: ownDarkColorName,
-                      inherited: true,
-                      priorBoxedOrCollapsibleAncestor:
-                          parentIsBoxedOrCollapsible,
-                  },
-        };
-    }
-
-    return {
-        value: dependencyValues[ownDarkColorName],
-        source: {
-            authored: false,
-            colorName: ownDarkColorName,
-            inherited: false,
-            priorBoxedOrCollapsibleAncestor: false,
-        },
-    };
-}
-
-function shouldEmitSectionTitleColorDiagnostic({ source }) {
-    if (!source?.authored) {
-        return false;
-    }
-
-    if (!source.inherited) {
-        return true;
-    }
-
-    // Once an inherited authored color has already passed through a boxed or
-    // collapsible ancestor, that earlier heading bar owns the diagnostic.
-    return !source.priorBoxedOrCollapsibleAncestor;
-}
-
-/**
- * Appends a contrast accessibility diagnostic when an explicitly authored
- * heading background color fails WCAG AA (4.5:1) against the heading text.
- *
- * Only fires when `authorSet` is true (i.e. the author overrode the default)
- * and the color is a parseable concrete value (CSS variable references like
- * `var(--mainGray)` return `null` from `compositedContrastRatio` and are
- * silently skipped).
- *
- * @param {object} params
- * @param {object[]} params.diagnostics - Accumulator array.
- * @param {boolean} params.authorSet - Whether the author explicitly set the attribute.
- * @param {string} params.colorValue - The resolved color string.
- * @param {string} params.colorName - Attribute name for the diagnostic message.
- * @param {string} params.textColor - Heading text color (#000000 or #ffffff).
- * @param {string} params.canvasColor - Viewer canvas color for the current theme.
- * @param {string} [params.modeSuffix] - Optional mode suffix for the message.
- */
-function addSectionTitleColorContrastDiagnostic({
-    diagnostics,
-    authorSet,
-    colorValue,
-    colorName,
-    textColor,
-    canvasColor,
-    modeSuffix = "",
-}) {
-    if (!authorSet || !colorValue) {
-        return;
-    }
-
-    const ratio = compositedContrastRatio({
-        foreground: textColor,
-        canvas: canvasColor,
-        background: colorValue,
-    });
-    if (ratio !== null && ratio < TEXT_CONTRAST_THRESHOLD) {
-        diagnostics.push({
-            type: "accessibility",
-            level: 1,
-            message: `${colorName} has insufficient contrast for the section heading text${modeSuffix} (${ratio.toFixed(2)}:1; requires at least ${TEXT_CONTRAST_THRESHOLD}:1).`,
-        });
-    }
-}
+import {
+    CANVAS_DARK_MODE_COLOR,
+    CANVAS_LIGHT_MODE_COLOR,
+    sectionTitleStateKeys,
+    titleStateKeyFromCredit,
+    resolveSectionTitleLightColorSpec,
+    resolveSectionTitleDarkColorSpec,
+    shouldEmitSectionTitleColorDiagnostic,
+    addSectionTitleColorContrastDiagnostic,
+} from "../../utils/sectionTitleColors";
 
 export class SectioningComponent extends BlockComponent {
     constructor(args) {
